Cleaning up MembersLeaveButton and adjacent code
