+++ desktopApp/reactApp/src/pages/ConsortiumDetails/Members/Members.tsx
@@ -11,7 +11,7 @@ interface MembersProps {
 }
 
 export function Members({ members, activeMembers, readyMembers, leader }: MembersProps) {
-    const { memberList, setMemberActive, setMemberReady, consortiumId } = useMembers({ members, activeMembers, readyMembers, leader });
+    const { memberList, setMemberActive, setMemberReady, handleLeave } = useMembers({ members, activeMembers, readyMembers, leader });
 
-    return <MembersDisplay memberList={memberList} setMemberActive={setMemberActive} setMemberReady={setMemberReady} consortiumId={consortiumId}  />;
+    return <MembersDisplay memberList={memberList} setMemberActive={setMemberActive} setMemberReady={setMemberReady} handleLeave={handleLeave} />;
 }
